+++ apps/e2e/tests/authentication/auth.po.ts
@@ -64,7 +64,7 @@ export class AuthPageObject {
   createRandomEmail() {
     const value = Math.random() * 10000000000;
 
-    return `${value.toFixed(0)}@makerkit.dev`;
+    return `${value.toFixed(0)}@getieltsy.dev`;
   }
 
   async signUpFlow(path: string) {